intask_api/requirements.txt
Django==3.0.6
celery==4.4.2
djangorestframework==3.11.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
redis==3.5.2
